Re-architect search algorithm and it's interface (#57)
This PR is mostly refactoring, but it includes a pretty large number of structural changes. The overall aim is to better match the program architecture to the documented design, increase consistency, better encapsulate the code, and improve readability.

I changed the search algorithm abstraction by:

changing the CostFunction from a class with internal state to a member function
changing the Simulator from an abstract method to a member function
storing the simulator state in member variable rather than locals. The purpose of this is to allow the _should_stop() method (Convergence Detector in the docs) to access any properties of the current state if the algorithm
Along with these changes to the search algorithm architecture, I decided to change its test. Previously the test simply counted five iterations. I created a new test that is basically a really shitty O(n^2) search for the largest element in a list. I designed it to fully use our interface, basically the simulate step compares the first two elements and marks the smaller one as not the largest. The candidate generate then removes this element when it generates the next iteration, and the convergence detector looks for there just being two elements left, when the largest over all candidate will be marked with the lower cost.

The other smaller changes I made are:

creating a liberate_simulator which calls run_liberate, and I pulled the candidate list generation out into there, and pulled the ldb parsing into there
changed the grapher to use an already parsed ldb rather than parsing twice
Aliased the top level Group object in the parser to a more readable LibertyResult
Moved the noop cost function to its own file

* Abstract Group to LibertyResult outside parser

* Move NoopCostFunction to separate file

* Convert cost functions to functions

* Renamed return value of run_liberate()

* make simulator function and add global search state for convergence detector

* update test to use new interface and change test to actual search

* update single param sweep and move liberate simulator to liberate.py

* Use results from simulation rather than parsing twice

* Fix typing and remove unused code

* MFix missing return types

* Remove unneccessary lint waiver

* Add leading underscore for private members